Module lsynth::c_compatible

Provides C compatible functions for working with this library as a DLL.

Functions

ls_amplitude_slide⚠

Sends an AmplitudeSlide command to the given channel.

ls_force_set_amplitude⚠

Sends a ForceSetAmplitude command to the given channel.

ls_force_set_panning⚠

Sends a ForceSetPanning command to the given channel.

ls_frequency_slide⚠

Sends a FrequencySlide command to the given channel.

ls_generate⚠

Generates audio with the provided chip.

ls_get_tick_frames⚠

Returns the number of samples that are in a single tick.

ls_init

Initiates a new LSynth chip

ls_panning_slide⚠

Sends a PanningSlide command to the given channel.

ls_send_command⚠

Inserts a command into the provided chip_state

ls_set_amplitude⚠

Sends a SetAmplitude command to the given channel.

ls_set_custom_waveform⚠

Sends a SetCustomWaveform command to the given channel.

ls_set_frequency⚠

Sends a SetFrequency command to the given channel.

ls_set_panning⚠

Sends a SetPanning command to the given channel.

ls_set_phase⚠

Sends a SetPhase command to the given channel.

ls_set_waveform⚠

Sends a SetWaveform command to the given channel.